JWT added for authentication and authorization in postman for our api

+++ controllers/api/v1/post_api.js
@@ -22,29 +22,23 @@ module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
 
-    //if (post.user == req.user.id) {
-    post.remove();
+    if (post.user == req.user.id) {
+      post.remove();
 
-    await Comment.deleteMany({ post: req.params.id });
+      await Comment.deleteMany({ post: req.params.id });
 
-    /*if (req.xhr) {
-        return res.status(200).json({
-          data: {
-            post_id: req.params.id
-          },
-          message: "Post deleted!"
-        });
-      }*/
+      // req.flash("success", "Post and associated comments deleted!");
 
-    // req.flash("success", "Post and associated comments deleted!");
-
-    return res.json(200, {
-      message: "Post and associated comments deleted successfully",
-    });
-    /*} else {
-      req.flash("error", "You cannot delete this post!");
-      return res.redirect("back");
-    }*/
+      return res.json(200, {
+        message: "Post and associated comments deleted successfully",
+      });
+    } else {
+      return res.json(401, {
+        message: "You cannot delete this post",
+      });
+      //req.flash("error", "You cannot delete this post!");
+      //return res.redirect("back");
+    }
   } catch (err) {
     //req.flash("error", err);
     //return res.redirect("back");
